fix: moved content in navData
